requirements.txt
pycryptodome
plotly
bcrypt
matplotlib